move regioninformer into pkg
